Use a base class for dataflux checkpoint maanger.
Create a base class for the dataflux implementation of a lightning checkpoint manager. Create parent classes for each CheckpointIO and AsyncCheckpointIO base classes.
